Don't list unpublished questions on /polls
We've also started adding view tests.

Note that we can test whether the contents of a query set match our
expectations by passing strings that ought to equal the return value of
a query result's `__repr__` method. I think this is pretty neat.

I also like the fact that we can test both the markup, and the
template's context values, as both are exposed by the response object.