Run the app and database in Docker containers
Note that the Dockerfile is based on Debian slim, rather than Alpine
Linux. I would have used Alpine, but for some reason that I've not
managed to identify I was unable to make a TCP connection to Django's
server. No matter how many times I rebuilt it and fiddled with the port
settings I couldn't get Alpine to work. Switching to slim and rebuilding
the image resolved the problem immediately.